// backend/src/prisma/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly authorId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Post: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'];
      readonly authorId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['output'];
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly published: CodecTypes['pg/bool@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly user: {
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly post: {
      readonly authorId: CodecTypes['pg/text@1']['input'];
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly published: CodecTypes['pg/bool@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly user: {
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;